LayoutTests/storage/indexeddb/mozilla/global-data.html

<!DOCTYPE html>
<!--
  original test: http://mxr.mozilla.org/mozilla2.0/source/dom/indexedDB/test/test_global_data.html
  license of original test:
    " Any copyright is dedicated to the Public Domain.
      http://creativecommons.org/publicdomain/zero/1.0/ "
-->
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script src="../resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test IndexedDB's opening DB more than once");
if (window.layoutTestController)
    layoutTestController.waitUntilDone();

function test()
{
    indexedDB = evalAndLog("indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;");
    shouldBeFalse("indexedDB == null");
    IDBDatabaseException = evalAndLog("IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;");
    shouldBeFalse("IDBDatabaseException == null");

    name = window.location.pathname;
    description = "My Test Database";
    request = evalAndLog("indexedDB.open(name, description)");
    request.onsuccess = openSuccess;
    request.onerror = unexpectedErrorCallback;
}

function openSuccess()
{
    db1 = evalAndLog("db1 = event.target.result");

    request = evalAndLog("request = db1.setVersion('1')");
    request.onsuccess = cleanDatabase;
    request.onerror = unexpectedErrorCallback;
}

function cleanDatabase()
{
    deleteAllObjectStores(db1);

    objectStoreName = "Objects";
    objectStoreOptions = { keyPath: 'id', autoIncrement: true };
    evalAndLog("db1.createObjectStore(objectStoreName, objectStoreOptions);");

    request = evalAndLog("indexedDB.open(name, description);");
    request.onsuccess = open2Success;
    request.onerror = unexpectedErrorCallback;
}

function open2Success()
{
    db2 = evalAndLog("db2 = event.target.result");

    shouldBeTrue("db1 !== db2");
    shouldBe("db1.objectStoreNames.length", "1");
    shouldBe("db1.objectStoreNames.item(0)", "objectStoreName");
    shouldBe("db2.objectStoreNames.length", "1");
    shouldBe("db2.objectStoreNames.item(0)", "objectStoreName");

    objectStore1 = evalAndLog("objectStore1 = db1.transaction(objectStoreName).objectStore(objectStoreName);");
    shouldBe("objectStore1.name", "objectStoreName");
    shouldBe("objectStore1.keyPath", "objectStoreOptions.keyPath");

    objectStore2 = evalAndLog("objectStore2 = db2.transaction(objectStoreName).objectStore(objectStoreName);");
    shouldBeTrue("objectStore1 !== objectStore2");
    shouldBe("objectStore2.name", "objectStoreName");
    shouldBe("objectStore2.keyPath", "objectStoreOptions.keyPath");

    done();
}


test();

</script>
</body>
</html>
